+++ src/app/tips/career/page.mdx
@@ -1,7 +1,7 @@
 import { Footnote } from '@/components/atoms/Footnote';
 import { UlCol } from '@/components/atoms/UlCol';
 import { CAREER_PAGE_META } from '@/data/links';
-import { MDX } from '@/features/mdx';
+import { MDX } from '@/mdx-components';
 import { IncomeBarChart } from '@/features/career/components/IncomeBarChart';
 
 export const metadata = {
