+++ src/app/tips/compare/page.mdx
@@ -2,7 +2,7 @@ import { Footnote } from '@/components/atoms/Footnote';
 import { Tab, TabGroup, TabList, TabPanels, TabPanel } from '@/components/atoms/Tab';
 import { ENGLISH_REQUIREMENT_BY_EXAM } from '@/data/english-requirement';
 import { ExrateAnnotation } from '@/features/forex';
-import { MDX } from '@/features/mdx';
+import { MDX } from '@/mdx-components';
 import { CurrencySwitchControl, LDSSwitchControl } from '@/features/stats';
 import { CostTable } from '@/features/compare/components/CostTable';
 import { EnglishRequirementByExamTable } from '@/features/compare/components/EnglishRequirementByExamTable';
