+++ src/app/tips/internship/page.mdx
@@ -1,7 +1,7 @@
 import { Footnote } from '@/components/atoms/Footnote';
 import { UlCol } from '@/components/atoms/UlCol';
 import { INTERNSHIP_PAGE_META } from '@/data/links';
-import { MDX } from '@/features/mdx';
+import { MDX } from '@/mdx-components';
 import { InternSalaryPieChart } from '@/features/internship/components/InternSalaryPieChart';
 
 export const metadata = {
